requirements.txt
markitdown
markitdown[all]
reportlab
python-pptx